+++ R/char.diff.R
@@ -2,10 +2,38 @@
 #'
 #' @description Calculates the character difference from a discrete matrix
 #'
-#' @param matrix A discrete matrix or a list containing discrete characters.
+#' @param matrix A discrete matrix or a list containing discrete characters. The differences is calculated between the columns (usually characters). Use \code{t(matrix)} to calculate the differences between the rows.
+#' @param method The method to measure difference: \code{"hamming"} (default), \code{"gower"}, \code{"euclidean"}, \code{"ged"} or \code{"mord"}.
+#' @param translate \code{logical}, whether to translate the characters following the \emph{xyz} notation (\code{TRUE} - default; see details - Felsenstein XXX) or not (\code{FALSE}). Translation works for up to 26 tokens per character.
+#' @param special.tokens optional, a named \code{vector} of special tokens to be passed to \code{\link[base]{grep}} (make sure to protect the character with \code{"\\\\"}). By default \code{special.tokens <- c(missing = "\\\\?", inapplicable = "\\\\-", polymorphism = "\\\\&", uncertainty = "\\\\/")}.
+#' @param special.behaviours optional, a \code{list} of one or more functions for a special behaviour for \code{special.tokens}. See details.
 #' 
 #' @details
-#' The character difference is calculated as half the sum of the Gower distances between the characters.
+#' The different distances to calculate are:
+#' \itemize{
+#'      \item \code{"hamming"} The scaled hamming distance: the relative distance between each pairs of comparable characters (i.e. does not interpret the character token significance; the differences are non-ordered Fitch-like: 0-2 = 1).
+#'      \item \code{"gower"} The scaled gower distance: the absolute distance between each pairs of comparable characters (e.g. does interpret the character token significance; the differences are ordered and absolute: 0-2 = 2). The characters are translated following the \emph{xyz} notation (The first token is translated to 1, the second to 2, etc. - Felsenstein XXX).
+#'      \item \code{"euclidean"} The Euclidean distance: XXX
+#'      \item \code{"manhattan"} The Manhattan distance: XXX
+#'      \item \code{"ged"} The Generalised Euclidean Distance (Wills XXX): XXX
+#'      \item \code{"mord"} The maximum observable distance (Lloyd XXX): XXX
+#' }
+#' 
+#' When using \code{translate = TRUE}, the characters are translated following the \emph{xyz} notation where the first token is translated to 1, the second to 2, etc. For example, the character \code{0, 2, 1, 0} is translated to \code{1, 2, 3, 1}. In other words when \code{translate = TRUE}, the character tokens are not interpreted as numeric values. When using \code{translate = TRUE}, scaled metrics (i.e \code{"hamming"} and \code{"gower"}) are divide by \eqn{n-1} rather than \eqn{n} due to the first character always being equal to 1.
+#' 
+#' 
+#' 
+#' \code{special.behaviours} allows to generate a special rule for the \code{special.tokens}. The functions should can take the arguments \code{character, all_states} and should output a non 0 integer. By default, missing data returns all states, polymorphisms and uncertainties return all present states and inapplicable returns an extra state.
+#' 
+#' \itemize{
+#'      \item{code{missing = function(x,y) as.integer(y)}}
+#'      \item{code{inapplicable = function(x,y) as.integer(-1)}}
+#'      \item{code{polymorphism = function(x,y) as.integer(strsplit(x, split = "\\\\&")[[1]])}}
+#'      \item{code{uncertainty = function(x,y) as.integer(strsplit(x, split = "\\\\/")[[1]])}}
+#' }
+#'
+#' \code{x, y} as only inputs and should output a single value. Functions in the list should be named following the special token of concern (\code{x}). Elements of the list should be named as in \code{special.tokens}. For example, the special behaviour for the special token \code{"?"} can be coded as: \code{special.behaviours = list(missing = function(x, y) return(NA))} to make all comparisons containing the special token containing \code{"?"} return a difference of \code{NA}.
+#' 
 #' 
 #' @return
 #' A character difference value or a matrix of class \code{char.diff}
@@ -23,26 +51,68 @@
 #' @author Thomas Guillerme
 #' 
 #' @references
+#' Wills XXX
+#' Hamming XXX
+#' Lloyd XXX
+#' Felsenstein XXX
 #' Gower, J.C. 1966. Some distance properties of latent root and vector methods used in multivariate analysis. Biometrika 53:325-338.
 
-
-char.diff <- function (matrix)  {
+char.diff <- function(matrix, method = "hamming", translate = TRUE, special.tokens, special.behaviours)  {
 
     options(warn = -1)
 
     ## Sanitizing
     matrix_class <- check.class(matrix, c("matrix", "list"))
-    ## Method is Gower by default
-    method = "Gower"
-
     if(matrix_class == "list") {
-
         ## Check length
         check.length(matrix, 2, " must contain only two elements", errorif = FALSE)
 
         ## Convert into a matrix
         matrix <- matrix(c(unlist(matrix)), byrow = FALSE, ncol = 2)
     }
+
+    ## Method is Gower by default
+    avail_methods <- c("hamming", "gower", "manhattan", "euclidean", "ged", "mord")
+    check.method(method, avail_methods, msg = "method")
+
+    ## Special tokens
+    if(missing(special.tokens)) {
+        special.tokens <- character()
+    }
+    check.class(special.tokens, "character")
+    if(is.na(special.tokens["missing"])) {
+        special.tokens["missing"] <- "\\?"
+    }
+    if(is.na(special.tokens["inapplicable"])) {
+        special.tokens["inapplicable"] <- "\\-"
+    }
+    if(is.na(special.tokens["polymorphism"])) {
+        special.tokens["polymorphism"] <- "\\&"
+    }
+    if(is.na(special.tokens["uncertainty"])) {
+        special.tokens["uncertainty"] <- "\\/"
+    }
+
+    ## Special behaviours
+    if(missing(special.behaviours)) {
+        special.behaviours <- list()
+    }
+    check.class(special.behaviours, "list")
+    if(is.null(special.behaviours$missing)) {
+        special.behaviours$missing <- function(x,y) return(as.integer(y))
+    }
+    if(is.null(special.behaviours$inapplicable)) {
+        special.behaviours$inapplicable <- function(x,y) return(as.integer(-1))
+    }
+    if(is.null(special.behaviours$polymorphism)) {
+        special.behaviours$polymorphism <- function(x,y) return(as.integer(strsplit(x, split = "\\&")[[1]]))
+    }
+    if(is.null(special.behaviours$uncertainty)) {
+        special.behaviours$uncertainty <- function(x,y) return(as.integer(strsplit(x, split = "\\/")[[1]]))
+    }
+
+    ## translate
+    check.class(translate, "logical")
 
     ## Convert matrix (if not numeric)
     if(!all(apply(matrix, 2, class) == "numeric")) {
@@ -62,9 +132,19 @@ char.diff <- function (matrix)  {
     ## Setting the attributes
     attrs <- list(Size = N, Labels = dimnames(matrix)[[1L]], Diag = diag, Upper = upper, method = method, call = match.call(),  class = "dist")
 
+
     ## Calculating the gower distance
     #options(warn = -1) #TG: NA's get introduced. Don't care!
-    output <- as.matrix(.Call("C_char_diff", matrix, method, attrs))
+    switch(method,
+        hamming   = {output <- as.matrix(.Call("C_diff_hamming"  , matrix, method, as.integer(translate), attrs))},
+        gower     = {output <- as.matrix(.Call("C_diff_gower"    , matrix, method, as.integer(translate), attrs))},
+        euclidean = {output <- as.matrix(.Call("C_diff_euclidean", matrix, method, as.integer(translate), attrs))},
+        manhattan = {output <- as.matrix(dist(matrix, method = "manhattan"))},
+        ged       = {stop("ged not implemented yet");
+                     output <- as.matrix(.Call("C_diff_ged", matrix, method, attrs))},
+        mord      = {stop("mord not implemented yet");
+                     output <- as.matrix(.Call("C_diff_mord", matrix, method, attrs))},
+    )
     #options(warn = 0)
 
     ## Calculating the character difference
@@ -133,11 +213,11 @@ plot.char.diff <- function(x, ..., type = "matrix", legend = TRUE, legend.title 
     match_call <- match.call()
 
     ## Checking the input type
-    if(!any(class(matrix) == "matrix")) {
+    if(!is(matrix, "matrix")) {
         stop.call(match_call$matrix, " must be a matrix.")
     } else {
         ## If the input is just a matrix, calculate the characters differences
-        if(!any(class(matrix) == "char.diff")) {
+        if(!is(matrix, "char.diff")) {
             matrix <- char.diff(matrix)
         }
     }
@@ -183,7 +263,7 @@ plot.char.diff <- function(x, ..., type = "matrix", legend = TRUE, legend.title 
             check.class(main, "character")
             check.length(main, 1, " must be a single character string.", errorif = FALSE)
         }
-        if(class(legend) == "logical" & legend == TRUE) {
+        if(is(legend, "logical") && legend == TRUE) {
             legend <- c("Combined", "Individual")
         }
         if(missing(col)) {