+++ R/dispRity.metric.R
@@ -50,7 +50,9 @@
 #' The currently implemented dimension-level 2 metrics are:
 #' \itemize{
 #'   \item \code{ancestral.dist}: calculates the distance between each tip and node and their ancestral. This function needs either (1) \code{matrix}/\code{list} from \code{\link{nodes.coordinates}}; or a \code{tree} (\code{"phylo"}) and \code{full} (\code{"logical"}) argument to calculate the node coordinates for the direct descendants (\code{full = FALSE}) or all descendants down to the root (\code{full = TRUE}). NOTE: distance is calculated as \code{"euclidean"} by default, this can be changed using the \code{method} argument.
-#'
+#' 
+#'   \item \code{angles}: calculates the angles of the main axis of variation per dimension in a \code{matrix}. The angles are calculated using the least square algorithm from the \code{\link[stats]{lm}} function. The unit of the angle can be changed through the \code{unit} argument (either \code{"degree"} (default), \code{radian} or \code{slope}) and a base angle to measure the angle from can be passed through the \code{base} argument (by default \code{base = 0}, measuring the angle from the horizontal line (not that the \code{base} argument has to be passed in the same unit as \code{unit}).
+#' 
 #'   \item \code{centroids}: calculates the distance between each row and the centroid of the matrix (Lalibert'{e} 2010). This function can take an optional arguments \code{centroid} for defining the centroid (if missing (default), the centroid of the matrix is used). This argument can be either a subset of coordinates matching the matrix's dimensions (e.g. \code{c(0, 1, 2)} for a matrix with three columns) or a single value to be the coordinates of the centroid (e.g. \code{centroid = 0} will set the centroid coordinates to \code{c(0, 0, 0)} for a three dimensional matrix). NOTE: distance is calculated as \code{"euclidean"} by default, this can be changed using the \code{method} argument.
 #'
 #'   \item \code{displacements}: calculates the ratio between the distance to the centroid (see \code{centroids} above) and the distance from a reference (by default the origin of the space). The reference can be changed through the \code{reference} argument. NOTE: distance is calculated as \code{"euclidean"} by default, this can be changed using the \code{method} argument.
@@ -104,6 +106,12 @@
 #' ancestral.dist(dummy_matrix, nodes.coords = direct_anc_centroids)
 #' ## Calculating the distances from all the ancestral nodes
 #' ancestral.dist(dummy_matrix, nodes.coords = all_anc_centroids)
+#' 
+#' ## angles
+#' ## The angles in degrees of each axis
+#' angles(dummy_matrix)
+#' ## The angles in slope from the 1:1 slope (Beta = 1)
+#' angles(dummy_matrix, unit = "slope", base = 1)
 #'
 #' ## centroids
 #' ## Distances between each row and centroid of the matrix
@@ -215,6 +223,7 @@ dimension.level3.fun <- function(matrix, ...) {
 dimension.level2.fun <- function(matrix, ...) {
     cat("Dimension level 2 functions implemented in dispRity:\n")
     cat("?ancestral.dist\n")
+    cat("?angles\n")
     cat("?centroids\n")
     cat("?displacements\n")
     cat("?neighbours\n")
@@ -398,7 +407,7 @@ ancestral.dist <- function(matrix, nodes.coords, tree, full, method = "euclidean
         }
     }
     
-    switch(class(nodes.coords),
+    switch(class(nodes.coords)[1],
         matrix = {
             ## Converting both matrix and nodes.coords in lists
             matrix <- unlist(apply(matrix, 1, list), recursive = FALSE)
@@ -504,6 +513,41 @@ func.div <- function(matrix) {
     ## The FDiv metric
     return((sum(dist_centroid) - mean_dis_cent * (obs-1)) / ((sum(abs(dist_centroid - mean_dis_cent) + dist_centroid))/obs))
 }
+
+## Angles measurements
+angles <- function(matrix, unit = "degree", base = 0) {
+
+    ## Check the unit
+    all_methods <- c("degree", "radian", "slope")
+    check.method(unit, all_methods, "Angle unit")
+
+    ## Check the base
+    check.class(base, c("numeric", "integer"))
+
+    ## Generate the base angle (slope/radian/angle = 0)
+    base_angle <- seq_len(nrow(matrix))
+
+    ## Get all the slopes
+    slopes <- apply(matrix, 2, function(X, base_angle) lm(base_angle ~ X)$coefficients[[2]], base_angle = base_angle)
+
+    ## Convert the slopes
+    angles <- switch(unit,
+        #degree = {atan(slopes/(1 + base_slope * (base_slope + slopes))) * 180/pi},
+        degree = {atan(slopes) * 180/pi},
+        radian =  {atan(slopes)},
+        # radian = {atan(slopes/(1 + base_slope * (base_slope + slopes)))},
+        slope  = {slopes}
+        )
+
+    ## Add a base angle (if not 0)
+    if(base != 0) {
+        return(angles + base)
+    } else {
+        return(angles)
+    }
+}
+
+
 
 
 #' @title Nodes coordinates