+++ tests/testthat/test-boot.matrix.R
@@ -435,23 +435,9 @@ test_that("boot.matrix detects distance matrices", {
 })
 
 test_that("boot.matrix works with multiple trees AND probabilities", {
-
-    set.seed(444)
-    record <- paleotree::simFossilRecord(p = 0.1, q = 0.1, nruns = 1, nTotalTaxa = c(10,15), nExtant = c(10,15))
-    taxa <- paleotree::fossilRecord2fossilTaxa(record)
-    rangesCont <- paleotree::sampleRanges(taxa, r = 0.5)
-    cladogram <- paleotree::taxa2cladogram(taxa, plot = FALSE)
-    likFun <- paleotree::make_durationFreqCont(rangesCont)
-    srRes <- optim(paleotree::parInit(likFun), likFun, lower = paleotree::parLower(likFun), upper = paleotree::parUpper(likFun), method = "L-BFGS-B", control = list(maxit = 1000000))
-    sRate <- srRes[[1]][2]
-    divRate <- srRes[[1]][1]
-    tree <- paleotree::cal3TimePaleoPhy(cladogram, rangesCont, brRate = divRate, extRate = divRate, sampRate = sRate, ntrees = 2, plot = FALSE)
-    tree[[1]]$node.label <- tree[[2]]$node.label <- paste0("n", 1:Nnode(tree[[1]]))
-    ## Scale the trees to have the same most recent root age
-    tree[[1]]$root.time <- tree[[2]]$root.time <- tree[[2]]$root.time
-    ## Make the dummy data
-    set.seed(1)
-    data <- matrix(rnorm((Ntip(tree[[1]])+Nnode(tree[[1]]))*6), nrow = Ntip(tree[[1]])+Nnode(tree[[1]]), ncol = 6, dimnames = list(c(tree[[1]]$tip.label, tree[[1]]$node.label)))
+    load("paleotree_test_data.Rda")
+    tree <- paleotree_data$tree
+    data <- paleotree_data$data
 
     ## Works with a multiPhylo object
     time_slices_multree_normal <- chrono.subsets(data, tree, method = "continuous", time = 3, model = "proximity")
@@ -469,7 +455,7 @@ test_that("boot.matrix works with multiple trees AND probabilities", {
     test <- boot.matrix(time_slices_multree_normal, bootstraps = 7, rarefaction = TRUE)
     expect_is(test, "dispRity")
     expect_equal(unlist(lapply(test$subsets, lapply, length), use.name = FALSE),
-                 c(4, 14, 10, 35, 28, 21, 22, 77, 70, 63, 56, 49, 42, 35, 28, 21))
+                 c(6, 21, 10, 35, 28, 21, 20, 70, 63, 56, 49, 42, 35, 28, 21))
     warn <- capture_warning(test2 <- boot.matrix(time_slices_multree_normal, bootstraps = 7, boot.type = "single"))
     expect_equal(warn[[1]], "Multiple trees where used in time_slices_multree_normal. The 'boot.type' option is set to \"full\".")
     error <- capture_error(boot.matrix(time_slices_multree_normal, bootstraps = 7, prob = c("t1" = 0, "t12" = 0, "t11" = 0, "t8" = 0, "t7" = 0)))
@@ -479,11 +465,11 @@ test_that("boot.matrix works with multiple trees AND probabilities", {
     test <- boot.matrix(time_slices_multree_proba, bootstraps = 7)
     expect_is(test, "dispRity")
     expect_equal(sort(unlist(lapply(test$subsets, lapply, length), use.name = FALSE)),
-                 c(18, 21, 30, 35, 66, 77))
+                 c(18, 21, 42, 49, 60, 70))
     test <- boot.matrix(time_slices_multree_proba, bootstraps = 7, rarefaction = TRUE)
     expect_is(test, "dispRity")
     expect_equal(sort(unlist(lapply(test$subsets, lapply, length), use.name = FALSE)),
-                 c(18, 21, 21, 21, 28, 28, 30, 35, 35, 42, 49, 56, 63, 66, 70, 77))
+                 c(18, 21, 21, 21, 28, 28, 35, 35, 42, 42, 42, 49, 49, 56, 60, 63, 70))
 
     warn <- capture_warning(boot.matrix(time_slices_multree_proba, bootstraps = 7, boot.type = "single"))
     expect_equal(warn[[1]], "Multiple trees where used in time_slices_multree_proba. The 'boot.type' option is set to \"full\".")